Move FileTabs to shared folder for ACQ, Research file
