+++ source/frontend/src/features/properties/map/acquisition/AcquisitionContainer.tsx
@@ -2,7 +2,7 @@ import LoadingBackdrop from 'components/common/LoadingBackdrop';
 import { useMapSearch } from 'components/maps/hooks/useMapSearch';
 import { FileTypes } from 'constants/index';
 import { InventoryTabNames } from 'features/mapSideBar/property/InventoryTabs';
-import { FileTabType } from 'features/mapSideBar/tabs/FileTabs';
+import { FileTabType } from 'features/properties/map/shared/detail/FileTabs';
 import { FormikProps } from 'formik';
 import { useAcquisitionProvider } from 'hooks/repositories/useAcquisitionProvider';
 import useApiUserOverride from 'hooks/useApiUserOverride';
