+++ source/frontend/src/features/properties/map/acquisition/AcquisitionView.test.tsx
@@ -3,7 +3,7 @@ import MockAdapter from 'axios-mock-adapter';
 import { Claims } from 'constants/claims';
 import { FileTypes } from 'constants/index';
 import { InventoryTabNames } from 'features/mapSideBar/property/InventoryTabs';
-import { FileTabType } from 'features/mapSideBar/tabs/FileTabs';
+import { FileTabType } from 'features/properties/map/shared/detail/FileTabs';
 import { mockAcquisitionFileResponse } from 'mocks/acquisitionFiles.mock';
 import { mockLookups } from 'mocks/lookups.mock';
 import { mockNotesResponse } from 'mocks/noteResponses.mock';
