+++ source/frontend/src/features/properties/map/acquisition/detail/AcquisitionFileTabs.test.tsx
@@ -1,5 +1,5 @@
 import Claims from 'constants/claims';
-import { FileTabType } from 'features/mapSideBar/tabs/FileTabs';
+import { FileTabType } from 'features/properties/map/shared/detail/FileTabs';
 import { mockAcquisitionFileResponse } from 'mocks/acquisitionFiles.mock';
 import { act } from 'react-test-renderer';
 import { render, RenderOptions, userEvent, waitFor } from 'utils/test-utils';
