+++ source/frontend/src/features/properties/map/shared/detail/FileTabs.tsx
